add optional AllocationModeOptimized
in this mode after processing a packet we keep in memory the allocated slices
and we reuse them for new packets.
Slices are allocated in:

- recvPacket, we always request a slice with size = maxMsgLength
- when received a sshFxpReadPacket (downloads), we always request a slice with
size = maxTxPacket+9+4

The allocated slices have variable size.

Allocated slices are referenced to the request order id and are marked for reuse
after a request is served in maybeSendPackets.

The allocator is added to the packetManager struct and it is cleaned at the end
of the Serve() function.

This allocation mode is optional and disabled by default